made more robust the launch_ironpython_server. now it is checking the aedt_path and it is checking if connection to the client is successful before initializing it.
